requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.4
django-cors-headers==3.12.0
django-environ==0.8.1
django-restframework==0.0.1
django-tenant-schemas==1.11.0
djangorestframework==3.13.1
future==0.18.2
generics==3.5.0
ordered-set==4.1.0
psycopg2-binary==2.9.3
pytz==2022.1
serializers==0.2.4
six==1.16.0
sqlparse==0.4.2
